util: McPAT, adjusted interface code
In this commit the code that converts the Gem5 Stats to McPAT input is adjusted
to evaluate summations and other mathmatical expressions. This was causing
McPAT to improperly read the data out of the stats files and report an invalid
power estimation that did not match the activity reported by the Gem5 stats.

Change-Id: Iac550604c50de7e3eab304a2789d6f41eced79f4