mcpat_integration: Pseudo-op for ROI
This change package comes with an update to the mcpat parser to convert nan and
inf to 0 so that the mcpat has a stable output at all times. This commit also
has integrated a new pseudo-op that allows the programmer to place a ROI in
their code to skip any initialization, as the gem5+mcpat is slow.  # Please
enter the commit message for your changes. Lines starting

Change-Id: I4f419a5434e3c653af068bb99d58620e91be1a39