power-pred: Add base class for PowerPrediction Unit
This commit has a base class and a test child class for a power prediction
class uses a GlobalEvent to trigger every X cycles. The class can be extended
on by creating a derived class and overrwriting the lookup() method. A new
priority level was added so that the stat dump always happens before the power
predictor is invoked. The McPAT code must be moved from the stat dump code to a
new module for better code style.

Change-Id: Ic562e291252f135f50db39ecccc4c076200e39c3